{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

#

# modin-all
# pandas
# numpy
snowflake-snowpark-python

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
python-dotenv>=0.5.1
{% if cookiecutter.python_interpreter != 'python3' %}
# backwards compatibility
pathlib2
{% endif %}